squash and merge migrations file and rename token file
