+++ requirements.txt
@@ -1,4 +1,5 @@
 asgiref==3.8.1
+coverage==7.7.1
 Django==5.1.7
 django-filter==25.1
 djangorestframework==3.15.2
